Replaced malfunctioning web.config example
I followed the instructions on http://book.cakephp.org/2.0/en/installation/advanced-installation.html#url-rewrites-on-iis7-windows-hosts and even tried some other proposed solutions found on Google which all didn't work.
Then I was pointed to this lighthouse ticket http://cakephp.lighthouseapp.com/projects/43067/tickets/113-creating-webconfig-file-in-windows-for-ii7#ticket-113-5 where user Lee Chenhwa pointed out that he got it working with the solution proposed in the Stack Overflow question http://stackoverflow.com/questions/12497904/htaccess-and-cakephp-2-on-windows-iis7 created by the ticket creator. This working solution is an improved (read more flexible solution of the also solution http://www2.palomar.edu/pages/sphillips/cakephp-with-iis-7-rewrite-rules-in-a-sub-folder/ which was found in the mean time by the ticket creator.
I tested this solution on my Windows 2008 R2 with IIS 7.5 and it works. Even renamed my cakephp installation (which confirms the improved flexibility).