Update data-validation.rst
Added a new Tip to the 'Using Multiple Validation Rules' section.  This tip emphasizes the critical importance of knowing that the default values for 'allowEmpty' and 'required' are dependent upon the previous validation rules for the field.  There is short sentence above the new Tip regarding the default value dependency. I feel strongly that this short sentence should stay there in additional to the new Tip and that the two complement each other very well and only give a better understanding of this subtle, but critical to be aware of, feature.

Additionally, the new Tip points out a way to not only help avoid serious validation consequences if the field rules are changed without regard for the dependent default values, but it also is an optimization by doing so in that there is no need to explicitly set back either of those fields if the rules that use default values are listed before the rules that do not.