Proposal for some wording change about default.ctp
While reading through the Cookbook, I stumbled upon the statement that the default.ctp should be placed in app/View/Layouts/ but in the current version of CakePHP it's already there.
Also when you bake a new project.

Thus this proposal for changing the wording about that. I'm not fully content yet, so please comment.