Infinity looping bug
Whenever Cache lost the session data and Database has a copy of the expired data, ComboSession will not be able to destroy the expired session from Database and hence infinity looping happens (CakeSession rejected the data and attempted to destroy-start-destroy-start...). We solved this by modifying this code to make it deleting them from both Cache and Database regardless if Cache succeeded or not.