Official CakePHP Facebook group link added
As well as **Google Group** we should mention the **Facebook Group URL** in the CakePHP documentation. Facebook is now a popular social community and it's great to see the official cakephp group in Facebook. So we should add it so that people can keep in touch with CakePHP from their own facebook account.