Update retrieving-your-data.rst
Just a simple note. If no value is found it is not specified what 'find' will return hence the reader wouldn't know what to spect in that case.
I propose to add a note about it in the general find explanation.
As a general review I would propose to mention returned values in where ever function where is not mentioned. It would help a lot and it is just adding a single line in most cases