Fixed toc and link errors for epub build (Issue #1126). Add cover to epub.
* The epub build is derived from the html build in sphinx; however, some
sections, such as "Welcome" don't make sense to include in the epub.
Essentially, this means that the epub needs it's own toc file which
is what this commit adds (epub-contents.rst). Since there's no
epub_ type option to override the master_doc var, this has to be done
as an arg to sphinx-build.

* This commit also adds a cover to the epub so it looks more like a
real book. The cover isn't great but at least the framework will
be there for someone to add a real cover image.