Update part-two.rst
Because CakeRequest::is() has only been able to take an array as of CakePHP 2.4.0, earlier versions are expecting a string. So the example in edit() post is better added a comment to explain.
This can solve many new comers' problem.
And developer sometimes might use prior version to test due to production server limitation in real world practically.