clear up ambiguity on what to include
today in IRC there was confusion over the SimplePasswordHasher as in the docs the class does not appear to App:uses AppModel, only SimplePasswordHasher. To clarify for those following the tutorial who might not know what is going on, details were added to existing code and the App::uses('AppModel', 'Model') line was added.