pear-cakephp/cakephp no longer available
when adding this to require:

    "pear-cakephp/cakephp": ">=2.4.9"

Composer fails, unable to find cakephp. This *may* work after using pear to discover the new package, but should not be necessary to install with composer. After making this change, it worked perfectly.

I also updated the cake version to the latest version recommended by someone in IRC.