Add file name
First I was thinking if we do 
src/Model/Table/ArticlesTable.php
src/Controller/ArticlesController.php

In entities we do the same and...
src/Model/Entity/ArticlesEntity.php

But this is wrong.
It should be Article.php and this is not written in "Creating entity classes", so I included the file name to be easy to understand. I followed the pattern found in line 295.