Cache API missing parameters in some methods
The read and delete methods in the Cache API did not include the optional $config parameter in the documentation.  This can be confusing as it will read from 'default' automatically if nothing is specified in the $config parameter resulting in false being returned because the method could not find the specified key.