Start new bookmarks app tutorial.
The blog tutorial is getting a bit rusty. It spends a lot of time
covering really basic things without covering any of the cool things
that CakePHP can do.  I think this new tutorial still covers many of the
basic topics and covers some of the good features in Cake3.