Add back to contents button.
This button conditionally shows when you are scrolled below the page
contents and gives a quick way to get back to the page contents.
I didn't go back go the root nav as it can be pretty large, and I think
more people need an easy way to navigate around the current page.

Refs #1898