Update retrieving-data-and-resultsets.rst
The link to the query builder is quite important, since the query builder doc page extends this page. But when scrolling over this page, it's not very obvious. It just looks like the examples above are explained here and the link is pretty "hidden". This should help make the link (and it's importance) more clear.