Updated 'Using Custom Authentication Objects'
I've set the same example for both the 'app authentication object' and 'plugin authentication object'. I think that can be better understand in this way, having the 'Creating Custom Authentication Objects' and 'Using Custom Authentication Objects' with the same class name.

Anyway I've also added a note to clarify this.