Configuration Keys - log
Log is not one of the configuration keys that can be accessed using the methods for all the other configuration keys in this section via CakeEmail::log() or $Email->log();  It can only be accessed through the CakeEmail::config() method as CakeEmail::config(array('log' => true));  I moved the log to a not in the section about using CakeEmail::config.  Maybe in the long run the log method should be added to the CakeEmail class?