Case fix
This one was an evil one.
After deployment from OS X (NFS shared to vagrant) to Debian (Host) it did not find the ControllerAuthorize adapter. Reason was I had it in small case as it used to be in 2.x I think. Fixing the documentation here. Also it works with a string only in case you don't supply any configuration options. So I dropped it here.