Update saving-data.rst
I have to make this change for this code to work. Without the call to filter() any new associations are returned as a NULL element and the update fails. Without the toArray() call it doesn't work either. I'm guessing the code is OK and the problem is in the docs but asking for confirmation.