Update migrations.rst
This is not only my first CakePHP code change, but my first on GitHub as a whole as well, so please be gentle, ha.  

It is not clear via this documentation that, when using the command-line, if ":primary" is not appended to "id:uuid", an error will result.  Without specifying ":primary", the only way to get it to work is to manually modify the resulting migrations file, then proceed.  But this suggested note would have saved me (as a CakePHP newbie) some time searching, and judging by some Google searches, possibly some others as well.  I'm open to the wording, this is just a suggestion.