Update pagination.rst
The docs referring to the PaginatorComponent use an additional controller property than default. Instead of declaring the PaginatorComponent settings in $components = array('Paginator' => array( **SETTINGS** )). The user is told to add them to the $paginate property, which is NOT automatically picked up by Cake. In my opinion, due to cake's automatic behavior at times, this should be explicitly stated.