Better visibility on scope
This seems like an oversight to me. Deprecating scope will cause any application which rely on it who upgrade from 3.0 to 3.1 to allow anyone to login to the application without any warnings, notices or exceptions. So I think this should be much clearer in the book, to help people keep their applications secure.