Plugin::loadAll() doesn't work lazily
Imho it needs to be more clear that Plugin::loadAll() isn't really a completely lazy anymore in 3.x.   If you go through this tutorial there is a part that makes it sound like you can use loadAll() in place of load() and it will need no more configuration.  But actually you can't, and you have to specify routes is true, because it does not default to that with loadAll().