inflector.rst translation to portuguese
I've translated this file the best I could. For some reason, the first ".. note::" translates into a message warning the user that the portuguese version isn't available and I couldn't get rid of it.

Also it would be nice to translate "Note:" into "Nota:" but I couldn't figure out a way to do it.