Make overridable method signatures more copy/paste friendly.
Remove the type annotations that don't work in PHP5.5. This helps make
the documentation easier to copy and paste into application code. It
also helps make the signatures in the book better match the actual code
in the framework.

This is an alternate approach to solve the issue #3816 is aiming to solve.