Update form.rst
Documentation shall be more clear about validation sets and what they are. It is not too obvious in the validation section either, but I think it is better to mention it in FormHelper manual page, where we define which set of rules we want to use. 
At least here it will be apparent for developer what manual is talking about.