Add edge version warnings to 3.next
My thinking is that we can use the `3.next` branch in the docs as
a longer-lived branch that contains docs for the 'next' version of 3.x.

I'd like to deploy these changes to `http://book.cakephp.org/3.next/en`.
This will make it easier to host and link to docs for beta/rc versions
that contain dramatic changes from the current 3.x release.

When we merge `3.next` into `3.0` we'll need to update `config/all.py`
but as its a 1 line change I don' see this as being a big hurdle.