Remove unused component in the example
While reading the documentation i stumbled upon this example on how to use components and it confused me a little since the cookie component is mentioned and initialized but never used. So I figured we should either change the example to use the cookie component or just remove it. Since I could not think of a use case (in this context) for the cookie component I removed it.