Start adding new tutorial
In the discussion of #3806 and some additional discussion in slack,
building a new tutorial that combines the best elements of the blog
tutorial and the more advanced topics from the bookmarker into a single
tutorial was determined to be a good course of action. This is the start
of that new tutorial. Long term we'd be dropping the blog and bookmarker
tutorials from the documentation to streamline the experience for new
developers.