Added success
I'm always looking this up, and it appears to be missing from the documentation. I've not included all the tags present in the `ConsoleOutput` class, as many are the same colour, so it would duplication.

Let me know if you think it's worth linking to the relevant api section from the docs, and I can add that link.