Remove redundant login suggestion
At the end of the 'Add Login" section the tutorial rightfully suggests that "you should now be able to login." This is followed by a note about how to create a user with a hashed password if one was not previously created. This note is followed by another line about being able to log in that is redundant to the first. As a reader I would expect to be able to login after completing the proposed fix.