Clarify decimal type
This pull request aims to try and clarify the two different decimal parts, and what they are for. As precision refers to the whole integer part before the decimal point, and scale refers to the fractional part after the decimal point.

However this wasn't very clear, at least to me, from the way the documentation was worded.