fatal: ../src/Locale is outside repository
The [documentation](https://book.cakephp.org/4/en/appendices/4-0-upgrade-guide.html#use-the-upgrade-tool) directs the user to checkout the upgrade tool and then cd into the upgrade-folder and run the scripts from there. If the app is in a different repository the user will get the fatal error `../src/Locale is outside repository`.
The reason seems to be that the FileRenameCommand wants to use `git mv` if available, but it is not possible to do that to an outside folder.
So I propose to recommend the user to leave the upgrade folder again before attempting to run the tool.