Prepare indexing script for elasticsearch upgrade
I'm planning on upgrading elasticsearch for the book and plugin sites.
That will require a different index structure as es6 only allows one
type per index and recommends that the mapping be called _doc. This
updates the indexing script to allow both an es6 mode and the `--compat`
mode will work with our current elasticsearch version.