Update tags-and-users.rst
routes.php is different when built in CakePHP 4.0 - different units, object methods instead of static functions, and different variable names! It also comes loaded with a lot of comments, so either include those [though they are rather busy] or state they are removed - or neither and let the reader work it out!