Note virus-scanner can disrupt sending mail
https://discourse.cakephp.org/t/cant-send-mails-using-gmail-with-emailtransport-configured-as-described-in-cookbook-4/7786/9

Don't know how specific this is, only gmail, only Anti Virus Free (windows 10)? Seems to be php/tls issue?

Whatever the cause/combo's, at the moment it seems that you have take this in account when trying to get sending mail to work with gmail. Think it has to be mentioned.