Add a warning about executeCommand open files
When using `BaseCommand::executeCommand()` in a loop/iteration, it's
possible to reach a "Too many open files" (or similar) limitation
because a new instance of `ConsoleIo` is created if not supplied as
the 3rd argument, and that opens `stdin`, `stdout`, and `stderr`.

This adds a note about this potential edge-case usage.

See: https://github.com/cakephp/cakephp/issues/16259

It might be worth fixing, but for the time being, I think documenting
this behavior is a good idea.