Add build time language content cloning
To aid in the development of partial translations we can copy over
content from the english docs into the translation. This allows us to
have translations that use `:ref:`, `:doc:`, `:php:meth:`, `:php:class:`
links. In the past we would delete these links which cause build time
errors, but that only furthers the divergence in the docs that is
already naturally occurring.