Update paginator.rst
I had a very difficult time figuring out how to make the paginator compatible with my i18n and l10n needs. This is a minimalist approach. A possibly clearer option would be to have a section of the page specifically discussing how to do this and include an example.